requirements.txt
pandas==2.2.3
plotly-express==0.4.0
streamlit==1.42.1
requests==2.32.3